+++ packages/rx_bloc_cli/mason_templates/bricks/feature_onboarding/__brick__/lib/feature_email_change/blocs/email_change_bloc.dart
@@ -100,7 +100,7 @@ class EmailChangeBloc extends $EmailChangeBloc {
 
   Future<void> _navigateToNextStep(UserModel user) async {
     _goRouter.pop();
-    await _goRouter.push(ConfirmEmailRoute(user.email).location);
+    await _goRouter.go(ConfirmEmailRoute(user.email).location);
   }
 
   String? _validateEmail(Result<String> emailResult) {
