+++ packages/rx_bloc_cli/mason_templates/bricks/feature_onboarding/__brick__/lib/feature_onboarding/blocs/onboarding_bloc.dart
@@ -150,11 +150,11 @@ class OnboardingBloc extends $OnboardingBloc {
         return _router.go(const DashboardRoute().routeLocation);
       }
 
-        unawaited(_router.pushReplacement(OnboardingPhoneRoute().routeLocation));
+      _router.go(OnboardingPhoneRoute().routeLocation);
     }
-    unawaited(_router.pushReplacement(
+    _router.go(
       OnboardingEmailConfirmationRoute(user.email).routeLocation,
-    ));
+    );
   }
 
   @override
