+++ packages/rx_bloc_cli/mason_templates/bricks/feature_onboarding/__brick__/lib/feature_onboarding_email_confirmation/blocs/onboarding_email_confirmed_bloc.dart
@@ -80,7 +80,7 @@ class OnboardingEmailConfirmedBloc extends $OnboardingEmailConfirmedBloc {
         _$goToNextPageEvent.doOnData(
           (_) {
             if (_isOnboarding) {
-              _router.pushReplacement(const OnboardingPhoneRoute().routeLocation);
+              _router.go(const OnboardingPhoneRoute().routeLocation);
             } else {
               _router.go(const LoginRoute().routeLocation);
             }
