+++ packages/rx_bloc_cli/mason_templates/bricks/feature_onboarding/__brick__/lib/feature_onboarding_phone_confirmation/blocs/onboarding_phone_bloc.dart
@@ -136,9 +136,9 @@ class OnboardingPhoneBloc extends $OnboardingPhoneBloc {
           .whereSuccess()
           .doOnData((_) {
         if (_isOnboarding) {
-          _router.push(const OnboardingPhoneConfirmRoute().routeLocation);
+          _router.go(const OnboardingPhoneConfirmRoute().routeLocation);
         } else {
-         _router.push(const PhoneChangeConfirmRoute().routeLocation);
+          _router.go(const PhoneChangeConfirmRoute().routeLocation);
         }
       }).publish();
 }
