+++ packages/rx_bloc_cli/mason_templates/bricks/feature_showcase/__brick__/lib/feature_showcase/views/showcase_page.dart
@@ -40,7 +40,7 @@ class ShowcasePage extends StatelessWidget {
                     featureTitle: feature.title,
                     featureSubtitle: feature.subtitle,
                      onTap: () =>
-                        GoRouter.of(context).push(feature.route.routeLocation),
+                        GoRouter.of(context).go(feature.route.routeLocation),
                     icon: feature.icon,
                   );
                 },
