+++ packages/rx_bloc_cli/mason_templates/bricks/lib_auth/__brick__/lib/base/models/country_code_model.dart
@@ -1,3 +1,5 @@
+{{> licence.dart }}
+
 import 'package:json_annotation/json_annotation.dart';
 import 'package:widget_toolkit/models.dart';
