+++ packages/rx_bloc_cli/mason_templates/bricks/lib_auth/__brick__/lib/base/models/user_with_auth_token_model.dart
@@ -1,9 +1,4 @@
-// Copyright (c) 2023, Prime Holding JSC
-// https://www.primeholding.com
-//
-// Use of this source code is governed by an MIT-style
-// license that can be found in the LICENSE file or at
-// https://opensource.org/licenses/MIT.
+{{> licence.dart }}
 
 import 'package:copy_with_extension/copy_with_extension.dart';
 import 'package:equatable/equatable.dart';
