+++ packages/rx_bloc_cli/mason_templates/bricks/lib_mfa/__brick__/lib/feature_mfa/views/mfa_page.dart
@@ -110,7 +110,7 @@ class MfaPage extends StatelessWidget {
         retryButtonText: context.l10n.featureMfa.pinNotSetButtonText,
         retryCallback: (_) {
           context.pop();
-          GoRouter.of(context).push(
+          GoRouter.of(context).go(
             const SetPinRoute().routeLocation,
           );
         },
