+++ packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/lib/base/data_sources/local/pin_biometrics_local_data_source.dart
@@ -1,7 +1,7 @@
 {{> licence.dart }}
 
 import 'package:widget_toolkit_biometrics/widget_toolkit_biometrics.dart';
-import '../../base/data_sources/local/shared_preferences_instance.dart';
+import 'shared_preferences_instance.dart';
 
 /// You have to implement and provide a [BiometricsLocalDataSource], you can
 /// store the value of [_areBiometricsEnabled], for example in [SharedPreferences]
