+++ packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/lib/base/data_sources/local/pin_code_local_data_source.dart
@@ -10,7 +10,7 @@ class PinCodeLocalDataSource {
   final FlutterSecureStorage _storage;
   static const _storedPin = 'storedPin';
 
-/// Store the encrypted pin code in the secure storage
+  /// Store the encrypted pin code in the secure storage
   Future<void> storePin(String encryptedPin) async =>
       await _storage.write(key: _storedPin, value: encryptedPin);
