+++ packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/lib/base/models/pin_code/pin_code_update_request.dart
@@ -1,6 +1,5 @@
 {{> licence.dart }}
 
-
 import 'package:copy_with_extension/copy_with_extension.dart';
 import 'package:equatable/equatable.dart';
 import 'package:json_annotation/json_annotation.dart';
