+++ packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/lib/feature_pin_code/di/update_pin_page_with_dependencies.dart
@@ -4,7 +4,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:provider/provider.dart';
 
-import '../models/update_pin_model.dart';
+import '../../base/models/pin_code/update_pin_model.dart';
 import '../services/update_pin_code_service.dart';
 import '../views/update_pin_page.dart';
