--- packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/lib/feature_pin_code/models/create_pin_model.dart
@@ -1,21 +1,0 @@
-{{> licence.dart }}
-
-import '../../base/models/user_model.dart';
-
-sealed class CreatePinModel {}
-
-class CreatePinSetModel extends CreatePinModel {}
-
-class CreatePinConfirmModel extends CreatePinModel {
-  final String pinToConfirm;
-
-  CreatePinConfirmModel({
-    required this.pinToConfirm,
-  });
-}
-
-class CreatePinCompleteModel extends CreatePinModel {
-  final UserModel user;
-
-  CreatePinCompleteModel({required this.user});
-}
