--- packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/lib/feature_pin_code/models/pin_code_arguments.dart
@@ -1,15 +1,0 @@
-{{> licence.dart }}
-
-class PinCodeArguments {
-  const PinCodeArguments({
-    required this.title,
-    this.onReturn,
-    this.showBiometricsButton = false,
-    this.updateToken,
-  });
-
-  final String title;
-  final Function? onReturn;
-  final bool showBiometricsButton;
-  final String? updateToken;
-}
