--- packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/lib/feature_pin_code/models/update_pin_model.dart
@@ -1,29 +1,0 @@
-{{> licence.dart }}
-
-import '../../base/models/user_model.dart';
-
-sealed class UpdatePinModel {}
-
-class UpdatePinVerifyModel extends UpdatePinModel {}
-
-class UpdatePinSetModel extends UpdatePinModel {
-  final String token;
-
-  UpdatePinSetModel({required this.token});
-}
-
-class UpdatePinConfirmModel extends UpdatePinModel {
-  final String pinToConfirm;
-  final String token;
-
-  UpdatePinConfirmModel({
-    required this.pinToConfirm,
-    required this.token,
-  });
-}
-
-class UpdatePinCompleteModel extends UpdatePinModel {
-  final UserModel user;
-
-  UpdatePinCompleteModel({required this.user});
-}
