+++ packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/lib/feature_pin_code/repository/pin_biometrics_repository.dart
@@ -1,7 +1,7 @@
 {{> licence.dart }}
 import 'package:widget_toolkit_biometrics/widget_toolkit_biometrics.dart';
 
-import '../../feature_pin_code/data_source/pin_biometrics_local_data_source.dart';
+import '../../base/data_sources/local/pin_biometrics_local_data_source.dart';
 
 class PinBiometricsRepository {
   PinBiometricsRepository(this._pinBiometricsLocalDataSource);
