+++ packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/lib/feature_pin_code/services/create_pin_code_service.dart
@@ -1,11 +1,11 @@
 {{> licence.dart }}
 
 import 'package:widget_toolkit_pin/widget_toolkit_pin.dart';
+
 import '../../assets.dart';
 import '../../base/models/errors/error_model.dart';
-import '../models/create_pin_model.dart';
-
-import '../repository/pin_code_repository.dart';
+import '../../base/models/pin_code/create_pin_model.dart';
+import '../../base/repositories/pin_code_repository.dart';
 
 class CreatePinCodeService implements PinCodeService {
   CreatePinCodeService(
