+++ packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/lib/feature_pin_code/services/verify_pin_code_service.dart
@@ -2,7 +2,8 @@
 
 import 'package:widget_toolkit_pin/widget_toolkit_pin.dart';
 
-import '../repository/pin_code_repository.dart';
+import '../../base/repositories/pin_code_repository.dart';
+
 class VerifyPinCodeService implements PinCodeService {
   VerifyPinCodeService(this.pinCodeRepository);
