+++ packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/lib/feature_pin_code/views/set_pin_page.dart
@@ -10,9 +10,9 @@ import '../../app_extensions.dart';
 import '../../base/extensions/error_model_extensions.dart';
 import '../../base/extensions/error_model_translations.dart';
 
+import '../../base/models/pin_code/create_pin_model.dart';
 import '../../lib_auth/blocs/user_account_bloc.dart';
 import '../../lib_router/router.dart';
-import '../models/create_pin_model.dart';
 import '../services/create_pin_code_service.dart';
 
 class SetPinPage extends StatelessWidget {
