+++ packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/test/lib_pin_code/services/update_pin_code_service_test.dart
@@ -1,9 +1,9 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
+import 'package:{{project_name}}/base/models/pin_code/update_pin_model.dart';
 import 'package:{{project_name}}/base/models/user_model.dart';
-import 'package:{{project_name}}/feature_pin_code/models/update_pin_model.dart';
-import 'package:{{project_name}}/feature_pin_code/repository/pin_code_repository.dart';
+import 'package:{{project_name}}/base/repositories/pin_code_repository.dart';
 import 'package:{{project_name}}/feature_pin_code/services/update_pin_code_service.dart';
 
 import '../stubs.dart';
