+++ packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/test/lib_pin_code/views/mocks/pin_biometrics_local_datasource_mock.dart
@@ -1,6 +1,6 @@
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
-import 'package:{{project_name}}/feature_pin_code/data_source/pin_biometrics_local_data_source.dart';
+import 'package:{{project_name}}/base/data_sources/local/pin_biometrics_local_data_source.dart';
 
 import 'pin_biometrics_local_datasource_mock.mocks.dart';
