+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/base/app/{{project_name}}.dart
@@ -120,7 +120,7 @@ class __MyMaterialAppState extends State<_MyMaterialApp> {
         if (!router.routeInformationProvider.value.uri
             .toString()
             .contains(verifyPinRoute)) {
-          router.push(
+          router.go(
             verifyPinRoute,
           );
         }
